Added "Assessments" folder
This folder is intended to hold the important files (.CTL, .DAT, .PRJ, gmacsallout.OUT and gmacs.rep.) used in the past assessments. This will be useful to tack changes in inputs between versions of GMACS and new data. It allows to store for each year, each model that have been presented during the CPTs, etc...